Add files via upload
This is a program which uses a designed class called Num Days to store a value that represents a number of work hours and converts it to a number of days. For example, 8 hours would be converted to 1 day, 12 hours hours would be converted to .5 days, and 18 hours would be converted to 2.25 days.